Extract the methods for defining the path
This is to allow it to be used with Cluster
